Chore: Move lib/timing to lib/util/timing
