+++ lib/linter.js
@@ -24,7 +24,7 @@ const eslintScope = require("eslint-scope"),
     Traverser = require("./util/traverser"),
     createReportTranslator = require("./report-translator"),
     Rules = require("./rules"),
-    timing = require("./timing"),
+    timing = require("./util/timing"),
     astUtils = require("./ast-utils"),
     pkg = require("../package.json"),
     SourceCodeFixer = require("./util/source-code-fixer");
