+++ lib/util/timing.js
